# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for SMPN 1 Dumai website
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DeleteResponse {
  success: boolean;
  message: string;
}

export interface Berita {
  id: number;
  judul: string;
  konten: string;
  gambar?: string | null;
  kategori: string;
  penulis: string;
  slug: string;
  is_published: boolean;
  dilihat: number;
  created_at: string;
  updated_at: string;
}

export interface BeritaListResponse {
  data: Berita[];
  total: number;
  page: number;
  limit: number;
}

export interface CreateBeritaBody {
  judul: string;
  konten: string;
  gambar?: string | null;
  kategori: string;
  penulis: string;
  is_published: boolean;
}

export interface Guru {
  id: number;
  nama: string;
  nip?: string | null;
  mata_pelajaran: string;
  jabatan: string;
  foto?: string | null;
  bio?: string | null;
  is_active: boolean;
  created_at: string;
}

export interface CreateGuruBody {
  nama: string;
  nip?: string | null;
  mata_pelajaran: string;
  jabatan: string;
  foto?: string | null;
  bio?: string | null;
  is_active: boolean;
}

export interface Galeri {
  id: number;
  judul: string;
  gambar: string;
  kategori: string;
  deskripsi?: string | null;
  created_at: string;
}

export interface GaleriListResponse {
  data: Galeri[];
  total: number;
  page: number;
  limit: number;
}

export interface CreateGaleriBody {
  judul: string;
  gambar: string;
  kategori: string;
  deskripsi?: string | null;
}

export interface Pengumuman {
  id: number;
  judul: string;
  konten: string;
  prioritas: number;
  is_active: boolean;
  created_at: string;
}

export interface CreatePengumumanBody {
  judul: string;
  konten: string;
  prioritas: number;
  is_active: boolean;
}

export interface Prestasi {
  id: number;
  judul: string;
  deskripsi?: string | null;
  tanggal: string;
  tingkat: string;
  kategori: string;
  created_at: string;
}

export interface CreatePrestasiBody {
  judul: string;
  deskripsi?: string | null;
  tanggal: string;
  tingkat: string;
  kategori: string;
}

export interface Ekskul {
  id: number;
  nama: string;
  deskripsi?: string | null;
  gambar?: string | null;
  pembina?: string | null;
  jadwal?: string | null;
  is_active: boolean;
  created_at: string;
}

export interface CreateEkskulBody {
  nama: string;
  deskripsi?: string | null;
  gambar?: string | null;
  pembina?: string | null;
  jadwal?: string | null;
  is_active: boolean;
}

export interface Profil {
  id: number;
  nama_sekolah: string;
  npsn: string;
  alamat: string;
  telepon?: string | null;
  email?: string | null;
  website?: string | null;
  kepala_sekolah: string;
  foto_kepsek?: string | null;
  sambutan_kepsek?: string | null;
  visi: string;
  misi: string;
  sejarah?: string | null;
  akreditasi?: string | null;
  tahun_berdiri?: string | null;
  logo?: string | null;
}

export interface UpdateProfilBody {
  nama_sekolah: string;
  npsn: string;
  alamat: string;
  telepon?: string | null;
  email?: string | null;
  website?: string | null;
  kepala_sekolah: string;
  foto_kepsek?: string | null;
  sambutan_kepsek?: string | null;
  visi: string;
  misi: string;
  sejarah?: string | null;
  akreditasi?: string | null;
  tahun_berdiri?: string | null;
  logo?: string | null;
}

export interface Slider {
  id: number;
  judul: string;
  subjudul?: string | null;
  gambar: string;
  urutan: number;
  is_active: boolean;
}

export interface CreateSliderBody {
  judul: string;
  subjudul?: string | null;
  gambar: string;
  urutan: number;
  is_active: boolean;
}

export interface SpmbInfo {
  id: number;
  judul: string;
  konten: string;
  tanggal_buka?: string | null;
  tanggal_tutup?: string | null;
  kuota?: number | null;
  persyaratan?: string | null;
  link_pendaftaran?: string | null;
  is_active: boolean;
}

export interface UpdateSpmbInfoBody {
  judul: string;
  konten: string;
  tanggal_buka?: string | null;
  tanggal_tutup?: string | null;
  kuota?: number | null;
  persyaratan?: string | null;
  link_pendaftaran?: string | null;
  is_active: boolean;
}

export interface Statistik {
  total_siswa: number;
  total_guru: number;
  total_staff: number;
  total_kelas: number;
  pengunjung_hari_ini: number;
  total_pengunjung: number;
  total_prestasi: number;
  total_ekskul: number;
}

export interface PengunjungResponse {
  pengunjung_hari_ini: number;
  total_pengunjung: number;
}

export interface AdminLoginBody {
  username: string;
  password: string;
}

export interface AdminUser {
  id: number;
  username: string;
  nama: string;
  role: string;
}

export interface AdminLoginResponse {
  success: boolean;
  token: string;
  user: AdminUser;
}

export interface AdminDashboard {
  total_berita: number;
  total_guru: number;
  total_galeri: number;
  total_prestasi: number;
  total_pengunjung: number;
  pengunjung_hari_ini: number;
  berita_terbaru: Berita[];
}

export type ListBeritaParams = {
  page?: number;
  limit?: number;
  kategori?: string;
  search?: string;
};

export type ListGuruParams = {
  search?: string;
  mata_pelajaran?: string;
  sort?: string;
  order?: string;
};

export type ListGaleriParams = {
  kategori?: string;
  page?: number;
  limit?: number;
};

export type ListPrestasiParams = {
  kategori?: string;
};