# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for SMPN 1 Dumai website
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /berita:
    get:
      operationId: listBerita
      tags: [berita]
      summary: List all published articles
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 9
        - name: kategori
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BeritaListResponse"
    post:
      operationId: createBerita
      tags: [berita]
      summary: Create new article (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBeritaBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Berita"
...
  /berita/{id}:
    get:
      operationId: getBeritaById
      tags: [berita]
      summary: Get article by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Berita"
    put:
      operationId: updateBerita
      tags: [berita]
      summary: Update article (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBeritaBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Berita"
    delete:
      operationId: deleteBerita
      tags: [berita]
      summary: Delete article (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  # ─── GURU ────────────────────────────────────────────────────────────────────
...
  /galeri:
    get:
      operationId: listGaleri
      tags: [galeri]
      summary: List gallery photos
      parameters:
        - name: kategori
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Gallery photos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GaleriListResponse"
    post:
      operationId: createGaleri
      tags: [galeri]
      summary: Add gallery photo (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGaleriBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Galeri"
  /galeri/{id}:
    delete:
      operationId: deleteGaleri
      tags: [galeri]
      summary: Delete gallery photo (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  # ─── PENGUMUMAN ──────────────────────────────────────────────────────────────
...
  /ekskul:
    get:
      operationId: listEkskul
      tags: [ekskul]
      summary: List extracurricular activities
      responses:
        "200":
          description: List of extracurriculars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ekskul"
    post:
      operationId: createEkskul
      tags: [ekskul]
      summary: Create extracurricular (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEkskulBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ekskul"
  /ekskul/{id}:
    put:
      operationId: updateEkskul
      tags: [ekskul]
      summary: Update extracurricular (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEkskulBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ekskul"
    delete:
      operationId: deleteEkskul
      tags: [ekskul]
      summary: Delete extracurricular (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  # ─── PROFIL ──────────────────────────────────────────────────────────────────
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
...
  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin user
      responses:
        "200":
          description: Admin user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    # ── Berita ─────────────────────────────────────────────────────────────────
    Berita:
      type: object
      properties:
        id:
          type: integer
        judul:
          type: string
        konten:
          type: string
        gambar:
          type: string
          nullable: true
        kategori:
          type: string
        penulis:
          type: string
        slug:
          type: string
        is_published:
          type: boolean
        dilihat:
          type: integer
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - judul
        - konten
        - kategori
        - penulis
        - slug
        - is_published
        - dilihat
        - created_at
        - updated_at
    BeritaListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Berita"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    CreateBeritaBody:
      type: object
      properties:
        judul:
          type: string
        konten:
          type: string
        gambar:
          type: string
          nullable: true
        kategori:
          type: string
        penulis:
          type: string
        is_published:
          type: boolean
      required:
        - judul
        - konten
        - kategori
        - penulis
        - is_published

    # ── Guru ───────────────────────────────────────────────────────────────────
...
    Galeri:
      type: object
      properties:
        id:
          type: integer
        judul:
          type: string
        gambar:
          type: string
        kategori:
          type: string
        deskripsi:
          type: string
          nullable: true
        created_at:
          type: string
      required:
        - id
        - judul
        - gambar
        - kategori
        - created_at
    GaleriListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Galeri"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    CreateGaleriBody:
      type: object
      properties:
        judul:
          type: string
        gambar:
          type: string
        kategori:
          type: string
        deskripsi:
          type: string
          nullable: true
      required:
        - judul
        - gambar
        - kategori

    # ── Pengumuman ─────────────────────────────────────────────────────────────
...
    Ekskul:
      type: object
      properties:
        id:
          type: integer
        nama:
          type: string
        deskripsi:
          type: string
          nullable: true
        gambar:
          type: string
          nullable: true
        pembina:
          type: string
          nullable: true
        jadwal:
          type: string
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
      required:
        - id
        - nama
        - is_active
        - created_at
    CreateEkskulBody:
      type: object
      properties:
        nama:
          type: string
        deskripsi:
          type: string
          nullable: true
        gambar:
          type: string
          nullable: true
        pembina:
          type: string
          nullable: true
        jadwal:
          type: string
          nullable: true
        is_active:
          type: boolean
      required:
        - nama
        - is_active

    # ── Profil ─────────────────────────────────────────────────────────────────
...
    AdminLoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
        user:
          $ref: "#/components/schemas/AdminUser"
      required:
        - success
        - token
        - user
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        nama:
          type: string
        role:
          type: string
      required:
        - id
        - username
        - nama
        - role
...